refactor: Move useAutomaticGroups to feature-specific folder
- Move useAutomaticGroups.ts from /composables to /automatic-groups
- Update import paths to use relative imports
- Improve code organization with feature-based structure
- Keep feature-specific logic together for better maintainability

Benefits:
- Better feature encapsulation
- Easier to refactor/move entire feature
- Clear ownership and responsibility
- Follows feature-based architecture pattern
